api/requirements.txt
Django==2.1.7
djangorestframework==3.9.2
django-rest-swagger==2.2.0
psycopg2==2.7.5
django-cors-headers==2.5.2
pytz==2018.9
passlib==1.7.1
requests==2.21.0
